old-tests/submission/W3C/domhtml/HTMLTableCellElement03.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - TABLECELL</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-tdth-abbr" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<table summary="Table Summary">
<tbody><tr>
<th id="header-1">Employee Id</th>
<th abbr="hd1" align="center" axis="center" bgcolor="#00FFFF" char=":" charoff="1" colspan="1" headers="header-1" height="50" id="header-2" nowrap="nowrap" rowspan="1" scope="col" valign="middle" width="170">Employee Name</th>
<th>Position</th>
<th>Salary</th>
</tr>
<tr>
<td id="header-3">EMP0001</td>
<td abbr="hd2" align="center" axis="center" bgcolor="#FF0000" char=":" charoff="1" colspan="1" headers="header-3" height="50" id="header-4" nowrap="nowrap" rowspan="1" scope="col" valign="middle" width="175">Margaret Martin</td>
<td>Accountant</td>
<td>56,000</td>
</tr>
</tbody></table>


<script>
var domtest = async_test("The abbr, align, axis, height, and width IDL attributes of the td and th elements must reflect the respective content attributes of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vabbr;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("th");
      assert_equals(nodeList.length,4);
testNode = nodeList.item(1);
      vabbr = testNode.abbr;

      assert_equals(vabbr,"hd1");
       
    domtest.done();
}));

</script>
<p>The abbr, align, axis, height, and width IDL attributes of the td and th elements must reflect the respective content attributes of the same name.</p>
<div id="log"></div>
</body></html>
